demo/python/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.14.0
fsspec==2024.6.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
requests==2.32.3
sympy==1.12.1
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.12.1
urllib3==2.2.1
